+++ spec/features/auto_populating_home_address_from_postcode_search.rb
@@ -142,19 +142,19 @@ RSpec.feature "Teacher Early-Career Payments claims" do
       claim
     end
 
-    scenario "with Ordnance Survey API" do
+    scenario "with Ordnance Survey API data" do
       expect(claim.valid?(:submit)).to eq false
       visit claim_path(claim.policy.routing_name, "postcode-search")
 
       # - What is your home address
-      expect(page).to have_text(I18n.t("questions.home_address"))
+      expect(page).to have_text(I18n.t("questions.address.home.title"))
       expect(page).to have_link(href: claim_path(EarlyCareerPayments.routing_name, "address"))
 
       fill_in "Postcode", with: "SO16 9FX"
       click_on "Search"
 
       # - Select your home address
-      expect(page).to have_text(I18n.t("questions.home_address"))
+      expect(page).to have_text(I18n.t("questions.address.home.title"))
 
       choose "flat_11_millbrook_tower_windermere_avenue_southampton_so16_9fx"
       click_on "Continue"
@@ -165,7 +165,43 @@ RSpec.feature "Teacher Early-Career Payments claims" do
       expect(claim.postcode).to eql "SO16 9FX"
 
       # - What is your address
-      expect(page).not_to have_text(I18n.t("questions.address"))
+      expect(page).not_to have_text(I18n.t("questions.address.generic.title"))
+
+      # - Email address
+      expect(page).to have_text(I18n.t("questions.email_address"))
+    end
+
+    scenario "Claimant cannot find the correct address so chooses to manually enter address" do
+      expect(claim.valid?(:submit)).to eq false
+      visit claim_path(claim.policy.routing_name, "postcode-search")
+
+      # - What is your home address
+      expect(page).to have_text(I18n.t("questions.address.home.title"))
+      expect(page).to have_link(href: claim_path(EarlyCareerPayments.routing_name, "address"))
+
+      fill_in "Postcode", with: "SO16 9FX"
+      click_on "Search"
+
+      # - Select your home address
+      expect(page).to have_text(I18n.t("questions.address.home.title"))
+      expect(page).to have_text("Flat 11, Millbrook Tower, Windermere Avenue, Southampton, SO16 9FX")
+      expect(page).to have_link(href: claim_path(EarlyCareerPayments.routing_name, "address"))
+
+      click_link(I18n.t("questions.address.home.i_cannot_find"))
+
+      # - What is your address
+      expect(page).to have_text(I18n.t("questions.address.generic.title"))
+
+      fill_in :claim_address_line_1, with: "Penthouse Apartment, Millbrook Tower"
+      fill_in :claim_address_line_2, with: "Windermere Avenue"
+      fill_in "Town or city", with: "Southampton"
+      fill_in "Postcode", with: "SO16 9FX"
+      click_on "Continue"
+
+      expect(claim.reload.address_line_1).to eql("Penthouse Apartment, Millbrook Tower")
+      expect(claim.address_line_2).to eql("Windermere Avenue")
+      expect(claim.address_line_3).to eql("Southampton")
+      expect(claim.postcode).to eql("SO16 9FX")
 
       # - Email address
       expect(page).to have_text(I18n.t("questions.email_address"))
@@ -246,12 +282,12 @@ RSpec.feature "Teacher Early-Career Payments claims" do
       claim
     end
 
-    scenario "with Ordnance Survey API" do
+    scenario "with Ordnance Survey API data" do
       expect(claim.valid?(:submit)).to eq false
       visit claim_path(claim.policy.routing_name, "postcode-search")
 
       # - What is your home address
-      expect(page).to have_text(I18n.t("questions.home_address"))
+      expect(page).to have_text(I18n.t("questions.address.home.title"))
       expect(page).to have_link(href: claim_path(EarlyCareerPayments.routing_name, "address"))
 
       fill_in "Postcode", with: "SE13 7UN"
@@ -259,7 +295,7 @@ RSpec.feature "Teacher Early-Career Payments claims" do
       click_on "Search"
 
       # - Select your home address
-      expect(page).to have_text(I18n.t("questions.home_address"))
+      expect(page).to have_text(I18n.t("questions.address.home.title"))
 
       choose "38a_wearside_road_london_se13_7un"
       click_on "Continue"
@@ -270,7 +306,7 @@ RSpec.feature "Teacher Early-Career Payments claims" do
       expect(claim.postcode).to eql "SE13 7UN"
 
       # - What is your address
-      expect(page).not_to have_text(I18n.t("questions.address"))
+      expect(page).not_to have_text(I18n.t("questions.address.generic.title"))
 
       # - Email address
       expect(page).to have_text(I18n.t("questions.email_address"))
@@ -319,12 +355,12 @@ RSpec.feature "Teacher Early-Career Payments claims" do
       claim
     end
 
-    scenario "with Ordnance Survey API" do
+    scenario "when no results returned from the API display error message" do
       expect(claim.valid?(:submit)).to eq false
       visit claim_path(claim.policy.routing_name, "postcode-search")
 
       # - What is your home address
-      expect(page).to have_text(I18n.t("questions.home_address"))
+      expect(page).to have_text(I18n.t("questions.address.home.title"))
       expect(page).to have_link(href: claim_path(EarlyCareerPayments.routing_name, "address"))
 
       fill_in "Postcode", with: "SE13 7UN"
@@ -332,7 +368,7 @@ RSpec.feature "Teacher Early-Career Payments claims" do
       click_on "Search"
 
       # - Select your home address
-      expect(page).to have_text(I18n.t("questions.home_address"))
+      expect(page).to have_text(I18n.t("questions.address.home.title"))
       expect(page).to have_text("There is a problem")
       expect(page).to have_text("Postcode not found")