+++ spec/helpers/claims_helper_spec.rb
@@ -45,7 +45,7 @@ describe ClaimsHelper do
       it "returns an array of identity-related questions and answers for displaying to the user for review" do
         expected_answers = [
           [I18n.t("questions.name"), "Jo Bloggs", "name"],
-          [I18n.t("questions.address"), "Flat 1, 1 Test Road, Test Town, AB1 2CD", "address"],
+          [I18n.t("questions.address.generic.title"), "Flat 1, 1 Test Road, Test Town, AB1 2CD", "address"],
           [I18n.t("questions.date_of_birth"), "10 January 1980", "date-of-birth"],
           [I18n.t("questions.payroll_gender"), "Don’t know", "gender"],
           [I18n.t("questions.teacher_reference_number"), "1234567", "teacher-reference-number"],
@@ -73,7 +73,7 @@ describe ClaimsHelper do
 
         expected_answers = [
           [I18n.t("questions.name"), "Jo Bloggs", "name"],
-          [I18n.t("questions.address"), "Flat 1, 1 Test Road, Test Town, AB1 2CD", "address"],
+          [I18n.t("questions.address.generic.title"), "Flat 1, 1 Test Road, Test Town, AB1 2CD", "address"],
           [I18n.t("questions.date_of_birth"), nil, "date-of-birth"],
           [I18n.t("questions.payroll_gender"), "Don’t know", "gender"],
           [I18n.t("questions.teacher_reference_number"), "1234567", "teacher-reference-number"],
@@ -94,7 +94,7 @@ describe ClaimsHelper do
 
         expected_answers = [
           [I18n.t("questions.name"), "Jo Bloggs", "personal-details"],
-          [I18n.t("questions.address"), "Flat 1, 1 Test Road, Test Town, AB1 2CD", "address"],
+          [I18n.t("questions.address.generic.title"), "Flat 1, 1 Test Road, Test Town, AB1 2CD", "address"],
           [I18n.t("questions.date_of_birth"), "10 January 1980", "personal-details"],
           [I18n.t("questions.payroll_gender"), "Don’t know", "gender"],
           [I18n.t("questions.teacher_reference_number"), "1234567", "teacher-reference-number"],
@@ -112,7 +112,7 @@ describe ClaimsHelper do
 
         expected_answers = [
           [I18n.t("questions.name"), "Jo Bloggs", "personal-details"],
-          [I18n.t("questions.address"), "Flat 1, 1 Test Road, Test Town, AB1 2CD", "address"],
+          [I18n.t("questions.address.generic.title"), "Flat 1, 1 Test Road, Test Town, AB1 2CD", "address"],
           [I18n.t("questions.date_of_birth"), nil, "personal-details"],
           [I18n.t("questions.payroll_gender"), "Don’t know", "gender"],
           [I18n.t("questions.teacher_reference_number"), "1234567", "teacher-reference-number"],